Update file structure and names
