+++ internal/gateway/certificate_secret.yaml
