+++ internal/gateway/kyma_gateway.go
@@ -4,56 +4,103 @@ import (
 	"context"
 	_ "embed"
 	"fmt"
+	"github.com/kyma-project/api-gateway/apis/gateway/v1beta1"
 	"github.com/kyma-project/api-gateway/apis/operator/v1alpha1"
-	"k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"github.com/kyma-project/api-gateway/controllers"
 	ctrl "sigs.k8s.io/controller-runtime"
-
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 const (
-	kymaGatewayName      = "kyma-gateway"
-	kymaGatewayNamespace = "kyma-system"
+	nonGardenerDomainName = "local.kyma.dev"
 )
 
-//go:embed kyma_gateway.yaml
-var kymaGatewayManifest []byte
+// ReconcileKymaGateway reconciles the kyma-gateway and creates all required resources for the Gateway to fully work. It also adds a finalizer to
+// APIGateway CR and handles the deletion of the resources if the APIGateway CR is deleted.
+// Returns a Status object with the result of the reconciliation and an error if the reconciliation failed.
+func ReconcileKymaGateway(ctx context.Context, k8sClient client.Client, apiGatewayCR v1alpha1.APIGateway) controllers.Status {
 
-func reconcileKymaGateway(ctx context.Context, k8sClient client.Client, apiGatewayCR v1alpha1.APIGateway, domain string) error {
-	isEnabled := isKymaGatewayEnabled(apiGatewayCR)
-	ctrl.Log.Info("Reconciling Kyma gateway", "KymaGatewayEnabled", isEnabled)
+	if isKymaGatewayEnabled(apiGatewayCR) {
+		if !hasKymaGatewayFinalizer(apiGatewayCR) {
+			if err := addKymaGatewayFinalizer(ctx, k8sClient, &apiGatewayCR); err != nil {
+				return controllers.ErrorStatus(err, "Failed to add finalizer during Kyma Gateway reconciliation")
+			}
+		}
+	} else {
+		apiRuleExists, err := anyApiRuleExists(ctx, k8sClient)
+		if err != nil {
+			return controllers.ErrorStatus(err, "Error during evaluation of Kyma Gateway reconciliation")
+		}
 
-	templateValues := make(map[string]string)
-	templateValues["Name"] = kymaGatewayName
-	templateValues["Namespace"] = kymaGatewayNamespace
-	templateValues["Domain"] = domain
-	templateValues["CertificateSecretName"] = kymaGatewayCertSecretName
+		// In the future, we want to be more selective and block the deletion of the Kyma gateway only if it is actually
+		// used by an APIRule, since currently an APIRule and no Kyma gateway always result in an error status.
+		if apiRuleExists {
+			return controllers.WarningStatus(fmt.Errorf("kyma gateway deletion blocked by APIRules"), "Kyma Gateway cannot be disabled because APIRules exist.")
+		}
 
-	resource, err := createUnstructuredResource(kymaGatewayManifest, templateValues)
+		if !hasKymaGatewayFinalizer(apiGatewayCR) {
+			ctrl.Log.Info("Kyma Gateway is disabled and no finalizer exists, reconciliation is skipped.")
+			return controllers.SuccessfulStatus()
+		}
+	}
+
+	isGardenerCluster, err := runsOnGardnerCluster(ctx, k8sClient)
 	if err != nil {
+		return controllers.ErrorStatus(err, "Error during Kyma Gateway reconciliation")
+	}
+
+	var reconcileErr error
+	if isGardenerCluster {
+		reconcileErr = reconcileGardenerKymaGateway(ctx, k8sClient, apiGatewayCR)
+	} else {
+		reconcileErr = reconcileNonGardenerKymaGateway(ctx, k8sClient, apiGatewayCR)
+	}
+
+	if reconcileErr != nil {
+		return controllers.ErrorStatus(reconcileErr, "Error during Kyma Gateway reconciliation")
+	}
+
+	// Besides on disabling the Kyma gateway, we also need to remove the finalizer on APIGateway deletion to make sure we are not blocking the deletion of the CR.
+	if !isKymaGatewayEnabled(apiGatewayCR) || apiGatewayCR.IsInGracefulDeletion() {
+		if err := removeKymaGatewayFinalizer(ctx, k8sClient, &apiGatewayCR); err != nil {
+			return controllers.ErrorStatus(err, "Failed to remove finalizer during Kyma Gateway reconciliation")
+		}
+	}
+
+	return controllers.SuccessfulStatus()
+}
+
+func anyApiRuleExists(ctx context.Context, k8sClient client.Client) (bool, error) {
+	apiRuleList := v1beta1.APIRuleList{}
+	err := k8sClient.List(ctx, &apiRuleList)
+	if err != nil {
+		return false, err
+	}
+
+	return len(apiRuleList.Items) > 0, nil
+}
+
+func reconcileGardenerKymaGateway(ctx context.Context, k8sClient client.Client, apiGatewayCR v1alpha1.APIGateway) error {
+	domain, err := getGardenerDomain(ctx, k8sClient)
+	if err != nil {
+		return fmt.Errorf("failed to get Kyma gateway domain: %v", err)
+	}
+
+	if err := reconcileKymaGatewayDnsEntry(ctx, k8sClient, apiGatewayCR, domain); err != nil {
 		return err
 	}
 
-	if !isEnabled || apiGatewayCR.IsInGracefulDeletion() {
-		return deleteKymaGateway(ctx, k8sClient, resource)
+	if err := reconcileKymaGatewayCertificate(ctx, k8sClient, apiGatewayCR, domain); err != nil {
+		return err
 	}
 
-	return createOrUpdateResource(ctx, k8sClient, resource)
+	return reconcileKymaGateway(ctx, k8sClient, apiGatewayCR, domain)
 }
 
-func isKymaGatewayEnabled(cr v1alpha1.APIGateway) bool {
-	return cr.Spec.EnableKymaGateway != nil && *cr.Spec.EnableKymaGateway == true
-}
-
-func deleteKymaGateway(ctx context.Context, k8sClient client.Client, kymaGateway unstructured.Unstructured) error {
-	ctrl.Log.Info("Deleting Kyma gateway if it exists")
-	err := k8sClient.Delete(ctx, &kymaGateway)
-
-	if err != nil && !errors.IsNotFound(err) {
-		return fmt.Errorf("failed to delete Kyma gateway: %v", err)
+func reconcileNonGardenerKymaGateway(ctx context.Context, k8sClient client.Client, apiGatewayCR v1alpha1.APIGateway) error {
+	if err := reconcileNonGardenerCertificateSecret(ctx, k8sClient, apiGatewayCR); err != nil {
+		return err
 	}
 
-	ctrl.Log.Info("Successfully deleted Kyma gateway")
-	return nil
+	return reconcileKymaGateway(ctx, k8sClient, apiGatewayCR, nonGardenerDomainName)
 }
